apps/backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg[binary]>=3.1
pydantic>=2
pydantic-settings>=2
python-dotenv>=1.0
alembic>=1.13
openai>=1.40

pymupdf
python-multipart
